+++ Fuegomon.cs
@@ -8,7 +8,7 @@ namespace Proyecto_POO
 {
     class Fuegomon : Roquemon
     {
-        public Fuegomon(string nombre, string tipo, int vida, int ataque, int defensa, int velocidad, float critico) : base(nombre, tipo, vida, ataque, defensa, velocidad, critico)
+        public Fuegomon(string nombre, string tipo, int vida, int ataque, int defensa, int velocidad, int critico) : base(nombre, tipo, vida, ataque, defensa, velocidad, critico)
         {
             this.nombre = nombre;
             this.tipo = tipo;
@@ -22,7 +22,7 @@ namespace Proyecto_POO
         public override void ataqueEspecial()
         {
             //ataca y aumenta el critico del roquemon por 20 por 3 turnos
-            
+            //this.critico += 20;
         }
     }
 }
